去掉部分页面的alert输出 添加拦截器栈 检验session域中的是否有student jsp页面检验是否登录 跳转到登录界面

+++ listen/src/com/listen/service/StudentService.java
@@ -17,7 +17,7 @@ public interface StudentService {
 
     JSON initGradetest();
 
-    void initGradeCode(Student s,Integer score);
+    void initGradeCode(Student s, float score);
 
     void openNewCheckPoint(Student s, String checkcount);
